extracted theme to custom theme
